Implement multiple literal prefix optimization.
Given a regex that matches if any one of N alternates matches *and*
each alternate has a literal prefix, do a substring search to skip
ahead in the input to first matching literal prefix.

The obvious form of such a regex is `a|b|c`, but the optimization
is implemented on compiled programs, so it works even for regexes
like `a*b`.